combined_requirements.txt
accelerate==0.29.3
aioboto3==11.2.0
aiobotocore==2.5.0
aiodns==3.0.0
aiofiles==23.2.1
aiohttp==3.8.4
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
aleph-alpha-client==2.17.0
altair==5.3.0
# anaconda-navigator==2.0.3
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
applaunchservices==0.2.1
argh==0.26.2
asgiref==3.8.1
assemblyai==0.15.0
async-timeout==4.0.2
asynctest==0.13.0
atomicwrites==1.4.0
backoff==2.2.1
backports.weakref==1.0.post1
bcrypt==4.1.2
bitsandbytes==0.42.0
bkcharts==0.2
black==19.10b0
blis==0.7.9
boto==2.49.0
boto3==1.26.76
botocore==1.29.76
Bottleneck==1.3.2
brotlipy==0.7.0
build==1.2.1
cachetools==5.3.3
catalogue==2.0.8
certifi==2024.2.2
cfgv==3.3.1
charset-normalizer==3.3.2
chroma-bullet==2.2.0
chroma-hnswlib==0.7.3
chroma-migrate==0.0.7
chromadb==0.4.24
click==8.1.7
clickhouse-connect==0.6.6
# clyent==1.2.2 Error
colorama==0.4.6
coloredlogs==15.0.1
# conda==4.10.3 Error
conda-build==3.21.4 Error
# conda-verify==3.4.2 Error
confection==0.0.4
contextlib2==0.6.0.post1
coverage==7.2.5
cssselect==1.2.0
cycler==0.10.0
cymem==2.0.7
cytoolz==0.11.0
dataclasses-json==0.5.13
deeplake==3.6.12
Deprecated==1.2.14
dill==0.3.6
diskcache==5.6.1
distlib==0.3.6
dnspython==2.3.0
docker==6.1.0
duckdb==0.7.1
duckduckgo-search==2.9.3
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastapi==0.110.1
fastcache==1.1.0
ffmpeg==1.4
ffmpy==0.3.2
filelock==3.12.0
flake8==3.8.3
flask-wtf==1.2.1
flatbuffers==24.3.25
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.2
git-filter-repo==2.38.0
gitdb==4.0.10
GitPython==3.1.31
gmpy2==2.0.8
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud==0.34.0
googleapis-common-protos==1.63.0
gpt-index==0.4.15
gradio==3.50.0
gradio-client==0.6.1
graphlib-backport==1.1.0
grpcio==1.64.0
gspread==5.10.0
gspread-pandas==3.2.2
gTTS==2.3.1
h11==0.14.0
h5py==2.10.0
HeapDict==1.0.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
humbug==0.3.2
identify==2.5.24
importlib-metadata==7.0.0
importlib-resources==6.4.0
jdcal==1.4.1
jmespath==1.0.1
jupyter==1.0.0
kubernetes==29.0.0
langchain==0.0.186
langcodes==3.3.0
langsmith==0.0.12
llama-index==0.6.0a6
llvmlite==0.36.0
locket==0.2.1
loguru==0.7.0
lz4==4.3.3
markdown-it-py==3.0.0
marshmallow==3.20.1
mccabe==0.6.1
mdurl==0.1.2
# mkl-fft==1.3.0 Error
mkl-service==2.3.0 Error
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.2.1
multidict==6.0.4
multipledispatch==0.6.0
multiprocess==0.70.14
murmurhash==1.0.9
mypy-extensions==0.4.3
# navigator-updater==0.2.1 Error
nodeenv==1.7.0
numcodecs==0.11.0
numexpr==2.8.4
numpy==1.24.4
oauthlib==3.2.2
olefile==0.46
ollama==0.1.9
onnxruntime==1.16.3
openai==0.27.2
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.1
outcome==1.2.0
overrides==7.7.0
pandas==1.2.5
parso==0.7.0
pathos==0.3.0
pathspec==0.7.0
pathtools==0.1.2
pathy==0.10.1
patsy==0.5.1
pep8==1.7.1
Pillow==10.0.0
pinecone-client==2.2.1
pkginfo==1.7.0
platformdirs==3.5.0
playsound==1.2.2
ply==3.11
posthog==3.5.0
pox==0.3.2
ppft==1.7.6.6
pre-commit==3.3.1
preshed==3.0.8
protobuf==4.25.3
pulsar-client==3.5.0
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1-modules==0.3.0
pycares==4.3.0
pycosat==0.6.3
pycurl==7.43.0.6
pydantic==2.7.1
pydantic-core==2.18.1
pydub==0.25.1
pygments==2.17.2
PyJWT==2.8.0
PyMuPDF==1.24.1
# PyMuPDFb==1.24.1 Error
pyodbc===4.0.1
pypdf==3.12.2
pypdf2==3.0.1
PyPika==0.48.9
pyproject-hooks==1.0.0
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-benchmark==4.0.0
pytest-cov==4.0.0
pytest-integration==0.2.3
pytest-mock==3.10.0
python-dotenv==1.0.0
python-multipart==0.0.9
PyYAML==6.0
pyzmq==20.0.0
QDarkStyle==2.8.1
QtPy==1.9.0
readability-lxml==0.8.1
redis==4.5.4
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
ruff==0.3.7
s3transfer==0.6.1
safetensors==0.4.3
scikit-image==0.16.2
# selenium==4.9.0 Error
semantic-version==2.10.0
shellingham==1.5.4
simplegeneric==0.8.1
smart-open==6.3.0
smmap==5.0.0
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.10.4
SQLAlchemy==1.4.49
srsly==2.4.6
starlette==0.37.2
tables==3.6.1
tenacity==8.2.3
terminado==0.9.4
thinc==8.1.10
tiktoken==0.3.3
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.12.0
torch==2.2.2
tqdm==4.65.0
transformers==4.30.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.14.0
# typer==0.12.3 Error
typing-extensions==4.11.0
typing-inspect==0.9.0
unicodecsv==0.14.1
uritemplate==4.1.1
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.23.0
wasabi==0.10.1
watchfiles==0.21.0
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
widgetsnbextension==3.5.1
wsproto==1.2.0
wtforms==3.1.2
xlwings==0.24.1
xlwt==1.3.0
xmltodict==0.12.0
yarl==1.9.2
zict==2.0.0
zope.event==4.5.0
zstandard==0.22.0
blinker==1.8.2
docstring_parser==0.16
ffprobe==0.5
Flask==3.0.3
google-ai-generativelanguage==0.1.0
google-cloud-aiplatform==1.52.0
google-cloud-bigquery==3.23.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.1.0rc1
google-resumable-media==2.7.0
grpc-google-iam-v1==0.13.0
grpcio-status==1.62.2
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
proto-plus==1.23.0
pyasn1_modules==0.4.0
pydantic_core==2.18.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
shapely==2.0.4
six==1.16.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.18.2
